renamed dependency file and related imports
