+++ backend/app/dependencies/__init__.py
@@ -1,0 +1,2 @@
+# Package marker for app.dependencies
+
