--- backend/app/dependencies/auth_deps.py
@@ -1,7 +1,0 @@
-# Extracts current user from JWT token    
-# Checks if user is active, has proper roles or permissions
-# Raises proper HTTP errors if auth fails
-#
-#
-#
-# Implemented after crud/db operations
